Add Project 'Progressive growing of GANs' to solution
