--- examples/keras/Progressive growing of GANs/Progressive growing of GANs.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "Progressive growing of GANs", "Progressive growing of GANs\Progressive growing of GANs.pyproj", "{E2CF5042-D835-4161-A6A7-07FE74D8135C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E2CF5042-D835-4161-A6A7-07FE74D8135C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E2CF5042-D835-4161-A6A7-07FE74D8135C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
