+++ examples/keras/Progressive growing of GANs/misc.py
