+++ examples/keras/Progressive growing of GANs/model.py
